Write fixes for Cygwin
`make check` now appears to be passing so this fixes #26. JIT is mostly
working too, but `make check` under JIT isn't passing yet. See #27. The
JIT is therefore still disabled by default on Cygwin.

+++ blink/ssemov.c
@@ -233,7 +233,7 @@ static void MovsdVpsWps(P) {
              "z3B"   // res0 = Get.....Mem[force64bit](RexbRm)
              "r1i"   // res1 = 0
              "z4C",  // 128-bit PutReg
-             0);
+             (u64)0);
     }
   }
 }
